# api/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kubernetes RBAC Generator API",
        "description": "A REST API service for visualizing and analyzing Kubernetes Role-Based Access Control (RBAC) configurations. This service provides comprehensive insights into roles, bindings, and principals across your Kubernetes clusters.",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/djryanj/k8s-rbactory-backend",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production server"
        }
    ],
...
    "paths": {
...
        "/cluster/info": {
            "get": {
                "tags": [
                    "cluster"
                ],
                "summary": "Get cluster information",
                "description": "Returns information about the connected Kubernetes cluster including version, namespaces, and node count",
                "operationId": "getClusterInfo",
                "responses": {
                    "200": {
                        "description": "Cluster information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusterInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/namespaces": {
            "get": {
                "tags": [
                    "cluster"
                ],
                "summary": "List namespaces",
                "description": "Returns all namespaces in the cluster",
                "operationId": "listNamespaces",
                "responses": {
                    "200": {
                        "description": "Namespaces retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NamespaceList"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "schemas": {
...
            "ClusterInfo": {
                "type": "object",
                "required": [
                    "version",
                    "namespaces",
                    "nodeCount"
                ],
                "properties": {
                    "version": {
                        "type": "string",
                        "description": "Kubernetes cluster version",
                        "example": "v1.28.0"
                    },
                    "namespaces": {
                        "type": "array",
                        "description": "List of namespaces in the cluster",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "default",
                            "kube-system",
                            "kube-public"
                        ]
                    },
                    "nodeCount": {
                        "type": "integer",
                        "description": "Number of nodes in the cluster",
                        "example": 3
                    }
                }
            },
            "NamespaceList": {
                "type": "object",
                "required": [
                    "namespaces"
                ],
                "properties": {
                    "namespaces": {
                        "type": "array",
                        "description": "List of namespace names",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "default",
                            "kube-system"
                        ]
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid namespace"
                    },
                    "errors": {
                        "type": "array",
                        "description": "Detailed validation errors",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "type": "string",
                        "description": "Field name",
                        "example": "namespace"
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message",
                        "example": "invalid name format: must be a valid Kubernetes resource name"
                    }
                }
            }
        },
        "responses": {
...
            "InternalServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "example": {
                            "error": "Internal server error"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token for authentication (if implemented)"
            }
        }
    },
    "security": []
}